refactor: separate target-js and target-php
