+++ index.d.ts
@@ -1,2 +1,2 @@
 export * from './runtime/san'
-export { ToJSCompiler, ToPHPCompiler } from './dist/index'
+export * from './dist/index'
