+++ src/index.ts
@@ -1,2 +1,14 @@
-export { ToJSCompiler } from './compilers/to-js-compiler'
-export { ToPHPCompiler } from './compilers/to-php-compiler'
+export { ToJSCompiler } from './target-js'
+export { ToPHPCompiler } from './target-php'
+
+export { autoCloseTags } from './utils/element'
+export { getANodeProps, getANodePropByName } from './utils/anode'
+export { Emitter } from './utils/emitter'
+
+export { SanSourceFile } from './models/san-sourcefile'
+export { SanApp } from './models/san-app'
+export { SanProject } from './models/san-project'
+export { Compiler } from './models/compiler'
+export { Component, isComponentLoader, COMPONENT_RESERVED_MEMBERS } from './models/component'
+export { Expression } from './models/expression'
+export { ANode } from './models/anode'
