+++ src/parsers/normalize-component.ts
