+++ src/parsers/san-app-parser.ts
@@ -1,8 +1,8 @@
 import { getComponentClassIdentifier, isChildClassOf } from '../utils/ast-util'
 import { ComponentClassFinder } from './component-class-finder'
 import { CommonJS } from '../loaders/common-js'
-import { ts2js } from '../transpilers/ts2js'
-import { normalizeComponentClass } from '../transformers/normalize-component'
+import { ts2js } from '../target-js/compilers/ts2js'
+import { normalizeComponentClass } from './normalize-component'
 import { SanSourceFile } from '../models/san-sourcefile'
 import { Project, SourceFile, ClassDeclaration } from 'ts-morph'
 import { getDefaultConfigPath } from './tsconfig'
