+++ src/target-js/utils/fs.ts
@@ -1,0 +1,5 @@
+import { readFileSync } from 'fs'
+
+export function readStringSync (path: string) {
+    return readFileSync(path, 'utf8')
+}
