+++ src/target-php/transformers/refactor-filters-property.ts
@@ -1,0 +1,5 @@
+import { PropertyDeclaration } from 'ts-morph'
+
+export function refactorFiltersProperty (filters: PropertyDeclaration, sanssr = 'san-ssr') {
+    filters.setType(`import("${sanssr}").SanSSRFiltersDeclarations`)
+}
