+++ src/utils/anode.ts
@@ -1,0 +1,57 @@
+import { camelCase } from 'lodash'
+import { ANode, ANodeProp } from '../models/anode'
+
+/**
+* 获取 ANode props 数组中相应 name 的项
+*/
+export function getANodePropByName (aNode: ANode, name: string) {
+    const index = aNode.hotspot.props[name]
+    if (index != null) {
+        return aNode.props[index]
+    }
+}
+
+export function getANodeProps (aNode: ANode) {
+    return aNode.props.map(p => {
+        const prop = { ...p, name: camelCase(p.name) }
+        postProp(prop)
+        return prop
+    })
+}
+
+/**
+* 对属性信息进行处理
+* 对组件的 binds 或者特殊的属性（比如 input 的 checked）需要处理
+*
+* 扁平化：
+* 当 text 解析只有一项时，要么就是 string，要么就是 interp
+* interp 有可能是绑定到组件属性的表达式，不希望被 eval text 成 string
+* 所以这里做个处理，只有一项时直接抽出来
+*
+* bool属性：
+* 当绑定项没有值时，默认为true
+*
+* @param {Object} prop 属性对象
+*/
+function postProp (prop: ANodeProp) {
+    let expr = prop.expr
+
+    if (expr.type === 7) {
+        switch (expr.segs.length) {
+        case 0:
+            if (prop.raw == null) {
+                prop.expr = {
+                    type: 3,
+                    value: true
+                }
+            }
+            break
+
+        case 1:
+            expr = prop.expr = expr.segs[0]
+            if (expr.type === 5 && expr.filters.length === 0) {
+                prop.expr = expr.expr
+            }
+        }
+    }
+}
