+++ src/utils/element.ts
@@ -1,0 +1,3 @@
+export const autoCloseTags = new Set(
+    'area,base,br,col,embed,hr,img,input,keygen,param,source,track,wbr'.split(',')
+)
