--- src/utils/php-util.ts
@@ -1,8 +1,0 @@
-export const reservedNames = [/^list$/i]
-
-export function isReserved (name: string) {
-    for (const reserved of reservedNames) {
-        if (reserved.test(name)) return true
-    }
-    return false
-}
